refactor: split eulerEvc.tsx into per-wrapper files
Move each wrapper component into its own file under eulerEvc/ with
shared styled components and helpers extracted to common.tsx.
